+++ Boa.Constrictor.Example/Tests/ScreenplayPlaywrightTest.cs
@@ -1,0 +1,34 @@
+namespace Boa.Constrictor.Example;
+
+using System.Threading.Tasks;
+using Boa.Constrictor.Playwright.Abilities;
+using Boa.Constrictor.Playwright.Questions;
+using Boa.Constrictor.Playwright.Tasks;
+using Boa.Constrictor.Screenplay;
+using FluentAssertions;
+using NUnit.Framework;
+
+public class ScreenplayPlaywrightTest
+{
+    private IActor Actor;
+
+    [SetUp]
+    public async Task Setup()
+    {
+        Actor = new Actor(name: "Keith", logger: new ConsoleLogger());
+        Actor.Can(await BrowseTheWebSynchronously.UsingChromium());
+    }
+
+    [Test]
+    public async Task TestWikipediaSearch()
+    {
+        await Actor.AttemptsToAsync(OpenNewPage.ToUrl(MainPage.Url));
+        var valueAttribute = await Actor.AskingForAsync(Attribute.Of("[name='search']", "value"));
+        valueAttribute.Should().BeNullOrEmpty();
+
+        await Actor.AttemptsToAsync(Fill.ValueTo("[name='search']", "Giant panda"));
+
+        // Todo: add support to wait for async questions
+        // Actor.WaitsUntil(Text.Of("[id='firstHeading'] span"), IsEqualTo.Value("Giant panda"));
+    }
+}